requirements.txt
google-generativeai
pyaudio
python-dotenv
deepgram-sdk==3.4.0
requests